Run refresh materialized views with security definer
Materialized views can only be refreshed by the owner. By setting the `security definer` attribute
on the script invoking the refresh materialized view command, the script will be run as the user
who created the script.

Therefore, the script qgep_network.refresh_network_simple needs to be created with the same user
that owns the materialized views.