Fix setting usage_current and function_hiearchic symbology attrs
These attributes are derived from connected reaches based on rules.
So far we relied on their order within the FROM clause of the UPDATE.
According to the postgres docs this is unreliable and the query should
only return one single row for the update.

See https://www.postgresql.org/docs/10/sql-update.html

When a FROM clause is present, what essentially happens is that the target table is joined to the tables mentioned in the from_item list, and each output row of the join represents an update operation for the target table. When using FROM you should ensure that the join produces at most one output row for each row to be modified. In other words, a target row shouldn't join to more than one row from the other table(s). If it does, then only one of the join rows will be used to update the target row, but which one will be used is not readily predictable.

Fixes https://github.com/QGEP/QGEP/issues/611