adjust labeling for multiple wastewater_nodes
According to VSA-DSS, a wastewater structure with a leaping weir is represented with two wastewater nodes (see https://vsa.ch/wiki/regenueberlauf-mit-streichwehr/) This requires us to partition the input and output labels by NE.fk_wastewater_structure instead of RP.fk_wastewater_networkelement. See lines 349 and 358
Additionally, the enumeration of inputs and outputs is only necessary if there is more than one input/output reach with the addition of the CASE WHENs in line 331-337, the index of the cover label, input label and output label is only used if there is more than one entry.