Update vw_oo_overflow.yaml
It is preferable to use a PointZ layer for overflows. Otherwise overflows with no fk_overflow_to are not visible (this can happen then the functioning of a special structure is not exactly known). In order to visualise the topology the line geometry can be calculated on the QGIS side.