Introduce CurveToLine for azimuth
Before this commit, the order was based on the azimuth between the reach point and the first vertex. @urskaufmann noted that the azimuth calculation might be imprecise for arcs. With an additional CurveToLine, the arcs are split into linestrings before choosing the vertex.
Also, the symbology now checks out with NULL values for bottom_level